Changed workflows to reflect new start/stop/enable/disable API
git-svn-id: https://xbox/svnroot/Qorus/trunk@4590 e9028eec-4831-4325-a917-4dc279fcb47c

+++ js/collections/notifications.js
@@ -1,11 +1,46 @@
 define(function (require) {
-  var localstorage = require('localstorage'),
-      Model = require('models/notification'),
-      Collection;
+  var _          = require('underscore'),
+      Backbone   = require('backbone'),
+      Model      = require('models/notification'),
+      Dispatcher = require('qorus/dispatcher'),
+      Collection, notifications;
+  
+  // init localstorage
+  require('localstorage');
     
   Collection = Backbone.Collection.extend({
     model: Model,
     localStorage: new Backbone.LocalStorage('Notifications'),
+    
+    dispatch: function (obj, ev) {
+      var alert, alert_type, alert_event;
+
+      if (ev === 'system') {
+        if (obj.eventstr === 'SYSTEM_SHUTDOWN') {
+          this.clear();
+        }        
+      } else if (ev === 'session:changed') {
+        this.clear();
+      } else if (ev === 'alert') {
+        alert_event = obj.eventstr.split('_');
+        alert = obj.info;
+        alert_type = alert_event[1];
+        
+        if (alert_event[2] === 'RAISED') {
+          this.create({
+            id: "alert-" + alert.id,
+            group: 'alerts-' + alert_type,
+            title: alert.alert,
+            type: 'error',
+            description: alert.name,
+            url: sprintf("/system/alerts/%s/%s", alert_type, alert.id)
+          });          
+        } else if (alert_event[2] === 'CLEARED') {
+          alert = this.get("alert-"+ alert.id);
+          if (alert) alert.destroy();
+        }
+      }
+    },
     
     createGroupList: function () {
       var group_names = this.pluck('group'), 
@@ -17,18 +52,34 @@ define(function (require) {
         groups[group] = {
           name: group,
           notifications: _.invoke(ntfc, 'toJSON')
-        }
+        };
       });
       return groups;
     }, 
     
     clear: function (group) {
-      _.each(this.where({ group: group}), function (model) {
-        model.destroy({ silent: true });
-      });
+      // ongoing alerts can't be cleared manually
+      if (group === 'alerts-ongoing') return false;
+      
+      if (group) {
+        _.each(this.where({ group: group}), function (model) {
+          model.destroy({ silent: true });
+        });
+        this.trigger(sprintf('cleared:%s', group));
+      } else {
+        var groups = this.pluck('group');
+        _(this.models).each(function (model) {
+          model.destroy({ silent: true });
+        });
+        this.reset();
+      }
       this.trigger('sync');
     }
   });
+  
+  notifications = new Collection();
+  
+  notifications.listenTo(Dispatcher, 'system session:changed alert', this.dispatch);
   
   return new Collection();
 });
